+++ api/src/main/java/ru/lapshina/api/ProductDto.java
@@ -1,16 +1,17 @@
 package ru.lapshina.api;
 
 
+import java.math.BigDecimal;
 
 public class ProductDto {
     private  Long id;
     private String title;
-    private int cost;
+    private BigDecimal cost;
 
     public ProductDto() {
     }
 
-    public ProductDto(Long id, String title, int cost) {
+    public ProductDto(Long id, String title, BigDecimal cost) {
         this.id = id;
         this.title = title;
         this.cost = cost;
@@ -32,11 +33,11 @@ public class ProductDto {
         this.title = title;
     }
 
-    public int getCost() {
+    public BigDecimal getCost() {
         return cost;
     }
 
-    public void setCost(int cost) {
+    public void setCost(BigDecimal cost) {
         this.cost = cost;
     }
 }
